+++ CHANGELOG.md
@@ -30,3 +30,10 @@
 - Add documentation for contributing.
 - Upgrade dependencies.
 - **BREAKING** Change types for "using" helper (@ekosz via [#113](https://github.com/mike-marcacci/node-redlock/pull/114/)).
+
+## v5.0.0
+
+- Fix regression of retryCount: -1. (fixes #149)
+- Export RedlockAbortSignal type. (fixes #138)
+- Issue an improved error when passing non-integer durations. (fixes #120)
+- Upgrade dependencies.